requirements.txt
# 医学知识问答RAG系统 - 核心依赖

# ====================
# 核心框架
# ====================
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.10

# ====================
# 向量数据库
# ====================
pymilvus>=2.3.0
faiss-cpu>=1.7.4

# ====================
# 数据处理
# ====================
pyspark>=3.5.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyarrow>=14.0.0  # Parquet列式存储
fastparquet>=2023.10.0  # 备用Parquet引擎

# ====================
# 对象存储与数据库
# ====================
minio>=7.2.0  # MinIO对象存储
pymongo>=4.6.0  # MongoDB文档数据库

# ====================
# 医学数据处理
# ====================
biopython>=1.81
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ====================
# LLM相关
# ====================
openai>=1.0.0  # 用于兼容硅基流动API
httpx>=0.24.0
tiktoken>=0.5.0

# ====================
# Web界面
# ====================
gradio>=4.0.0
streamlit>=1.28.0
plotly>=5.17.0

# ====================
# 缓存
# ====================
redis>=5.0.0
rank-bm25>=0.2.2

# ====================
# 工具库
# ====================
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
colorlog>=6.7.0

# ====================
# 评估
# ====================
rouge-score>=0.1.2
nltk>=3.8.0

# ====================
# 监控和日志
# ====================
tensorboard>=2.14.0
wandb>=0.15.0
loguru>=0.7.0

# ====================
# 数据库
# ====================
datasets>=2.14.0
pyarrow>=14.0.0

# ====================
# 消息队列 - Kafka
# ====================
kafka-python>=2.0.2
confluent-kafka>=2.3.0

# ====================
# 任务调度 - Airflow
# ====================
apache-airflow>=2.7.0
apache-airflow-providers-apache-kafka>=1.2.0


